+++ commands/frases.py
@@ -1,6 +1,6 @@
 import discord
 import random
-from utils.cooldowns import add_cooldown, get_cooldown, remove_cooldown
+from utils.cooldowns import add_cooldown, get_cooldown
 from utils.format_text import load_txt_file, separate_authors
 
 FRASES = load_txt_file("frases.txt")