+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # Changelog
 
+## [1.2.0] - 2025-12-17
+
+- (1.2.0! Yupii!!)
+- Añadido comando /queopinas para preguntarle a Hepatitis B(ot) qué opina.
+- Añadido comando /forzarqueopinas para forzar a Hepatitis B(ot) a opinar algo.
+- Añadido comando /funfact para que Hepatitis B(ot) tire un facto jijoso.
+- Añadido comando /forzarfunfact para forzar un facto jijoso de Hepatitis B(ot).
+- Añadido comando /changelog para ver el changelog de Hepatitis B(ot).
+
 ## [1.1.3] - 2025-12-16
 
 - (ptm ya tres updates en el mismo dia).
 - Añadida clase Debug (thx Haziel).
-- Comando /forzarfrase ahora puede ser utilizado por todos (se supone que esto era para debuggear y nada más, pero dado que los admins también pueden usarlo y la mayoría en el server es admin, no vale la pena, así que ahora es para todos, pero tiene cooldown)
+- Comando /forzarfrase ahora puede ser utilizado por todos (se supone que esto era para debuggear y nada más, pero dado que los admins también pueden usarlo y la mayoría en el server es admin, no vale la pena, así que ahora es para todos, pero tiene cooldown).
 - Comandos /mensaje, /reaccion y /noreaccion ahora pueden ser utlizados por todos (sé que está feo gatekeepear comandos, pero pensé que iba a armar quilombo. Para evitar eso, los comandos tienen cooldown).
 - Corregida la descripción del comando /repohepatitis.
 - Reesctructuración del código para facilitar legibilidad.
-- (yap king, I know)
+- (yap king, I know).
 
 ## [1.1.2] - 2025-12-16
 
@@ -24,7 +33,6 @@
 
 - Añadido comando /mascota para invocar una mascota del canal mascotas.
 - Posibilidad de controlar a Hepatitis B(ot).
-
 
 ## [1.0.0] - 2025-12-15